Refactor POP component warning tests and relocate logic
Moved `generatePopComponentWarnings` tests from `waste-pops.test.js` to `validation-warnings.test.js` to align with the function's location. This ensures better modularity and keeps tests close to the implementation.
